Move test_get_adapter_from_container to test_set_output.py
